simpleSupervisor/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.36.0
pydantic>=2.10.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
ollama>=0.1.7 